// src/utils/types.ts
export type UnitProfileDetails = {
  id: string;
  company_id: number;
  title: string;
  first_name: string;
  last_name: string;
  phone_number: string;
  address: string;
  description: string;
  email: string;
  image: string;
  gender: string;
  date_of_birth: string;
  company_name: string;
  location: string;
  agent: string;

  mode_of_application: string;
  company_status: string;
  customer?: string;
  business_type: string;
  funding_purpose: string;
  trading_style: string;
  company_number: string;
  other_funding_purpose: string;
};